+++ core/autoload/global.gd
@@ -100,7 +100,7 @@ func set_editor_text(text: String, keep_carets: bool = true) -> void:
 			selection.o = Vector2i(_editor.get_selection_origin_line(index), _editor.get_selection_origin_column(index))
 			selection.c = Vector2i(_editor.get_caret_line(index), _editor.get_caret_column(index))
 			carets.append(selection)
-		scroll = Vector2(get_editor().get_h_scroll_bar().get_value(), get_editor().get_v_scroll_bar().get_value())
+		scroll = Vector2(_editor.scroll_horizontal, _editor.scroll_vertical)
 	_editor.text = text
 	if keep_carets:
 		for idx: int in carets.size():
@@ -108,8 +108,8 @@ func set_editor_text(text: String, keep_carets: bool = true) -> void:
 			if idx >= _editor.get_caret_count():
 				_editor.add_caret(0, 0)
 			_editor.select(selection.o.x, selection.o.y, selection.c.x, selection.c.y, idx)
-		get_editor().get_h_scroll_bar().set_value(scroll.x)
-		get_editor().get_v_scroll_bar().set_value(scroll.y)
+		_editor.scroll_horizontal = scroll.x
+		_editor.scroll_vertical = scroll.y
 
 
 ## Will disable the editor if [param disabled] is [code]true[/code].